[LLCL] Implement (much of) AsyncValue and ConcurrentAppendingVector.
This implements a large swath of the core AsyncValue class.  It is
still missing support for IndirectAsyncValue and error handling
though.

This also implements a ConcurrentAppendingVector class, which is a
simple but useful concurrent collection.  It could be made much
better, but suffices for now.

MODULAR_ORIG_COMMIT_REV_ID: 6afc81a6d611203c59ce8391a5a8cce362f36ca7