[Header boilerplate] Standardize the first line comment in files.
This removes the "super short description" that is inevitably
duplicated with either the optional per-file comment, or is obvious
based on the filename.  Just remove it, which also eliminates
needless copypasta issues that rise up: of course the cmake file is
the build for the directory it is in. duh.

MODULAR_ORIG_COMMIT_REV_ID: ba8d0edd053135f114f299caa3b7ed7a87e557ac