[Algorithms] Add a parallel for loop.
This adds the version that takes a per-element function, a completion
handler, and an arbitrary amount of state to capture.  We should also
eventually have one that completes a chain when done (instead of
calling a completion handler), but this is a starting point.

MODULAR_ORIG_COMMIT_REV_ID: 4bce49e95e66d35321fc11b9907deac93d2b7194