[WorkQueue] Soften wording about 'await' being bad.
'await' doesn't actually block the client thread, it donates it to
running work in a work queue.  As such, it is perfectly fine to
use it from a workqueue thread (particularly for CPU bound
workloads like a parallel MLIR compiler), and this enables more
convenient design patterns which don't *have* to be non-blocking
continuation passing style for correctness.

That said, it can still be a good idea to use that pattern,
particularly when working with accelerators, because you often have
a primary thread that is "discovering" more work to cram onto the
accelerator.  LLCL should support both styles of programming.

MODULAR_ORIG_COMMIT_REV_ID: da83a199e454acb0775ad771dfe0a6488af36540