[parallelForEachN] Build out the family of parallelForEachN algorithms.
This renames the uber function from `parallelForEachN` to
`parallelForEachNCustomCompletion` since it is the advanced interface
that few people will need to work with.

This layers on several more progressively-derived versions of this
that are easier to use and important for other more specialized
use-cases.  Each takes the same "n" and per-element function, but
handle completion in different ways:

- `parallelForEachNMarkReady` marks an AsyncValue as "ready" when done.
- `parallelForEachNChain` returns a Chain to the client and marks
  it as ready when done.
- `parallelForEachN` donates the client thread to running work
  until all the elements are done.

The last one in particular gives a familiar and composable
interface that covers the common case.  Because it doesn't return
to the caller until the elements are complete, it is safe for the
per-element function to capture references to the callers stack.

MODULAR_ORIG_COMMIT_REV_ID: 9ea47cd427ca4167c79e57ae36f75f0c9d342413