[RCRef] privatize addRef/dropRef to discourage use.
This marks the addRef/dropRef members on ReferenceCounted and AsyncValue
as private so random client code doesn't use it.  This encourages the
use of RCRef/AsyncValueRef which is much safer and leads to more
declarative code.

There are still unusual situations where low-level access to the
reference count is needed.  For this, the RCRef class provides some
"lowLevel" static methods that have access to it.

MODULAR_ORIG_COMMIT_REV_ID: 1ca449379873b6a895429cab0d0be00c801b34b1