[Algorithms.h] Revise andThen implementations on multiple values.
This changes the heteogenous version to take an explicit tuple
value, which makes it more consistent (lambda at the end).

This renames andThen(ArrayRef) to andThenCopying/andThenMoving
to reflect that one copies the elements out of the array and the
other moves out.

This also fixes a problem with the andThen(ArrayRef) implementation
where it didn't extend the lifetime of the array elements it acted
on, which could cause dangling pointer problems.

MODULAR_ORIG_COMMIT_REV_ID: 92589f1cc9ce1e9decefd500721a8f8c1e21490a