[AsyncValue] Add ability to move from 'constructed' to 'error' state.
This requires destroying the constructed payload in an AsyncValue.
This in turn requires reworking how the destructor function is managed
for ConcreteAsyncValue's in general. Instead of capturing the dtor for
ConcreteAsyncValue<T>, just capture the destructor pointer for T.
This is more flexible, and potentially more reusable for other clients.

Use this in TFLite support to simplify memory management.

MODULAR_ORIG_COMMIT_REV_ID: bca02eebfaae89a26bcc03cbcc3bbf24333d35e5