[Support] Introduce a new ErrorOrSupport class. (#649)
This class is like `ErrorOr<void>`, allowing APIs and values that are
either successful (in the logical result sense) or hold an Error that
describes the problem.  This adopts it in a couple of places (e.g.
the GenericML kernel and the TFLExecution subsystem) but not
pervasively throughout the codebase yet.

This required refactoring how LogicalResult works in MLIR.  We want
to be able to use its `success()` method, but we don't want to allow
conversion from failure() to ErrorOrSupport.  Solve this by splitting
`success()` out to returning a new bespoke type that differentiates the
two of them.

This is progress towards solving Issue #615

MODULAR_ORIG_COMMIT_REV_ID: 4129c08559c200d32ca6d33cbca6e53195280ae5