[PrimitiveRegistry] Switch decoding logic to ErrorOrSuccess, NFC. (#653)
[PrimitiveRegistry] Switch decoding logic to ErrorOrSuccess, NFC.

This changes the pattern for decoding BEF attributes from "taking
an Error by-reference and returning a ParseResult" to just returning
an ErrorOrSuccess, which directly models the contract.

This tightens up a bunch of logic and eliminates the pontential for
bugs where failure is returned but the error isn't filled in, or when
error is filled in, but success is returned.

MODULAR_ORIG_COMMIT_REV_ID: cfe3061d736a33c70ad269dba11056d8c7c24f64