[GenericML] Remove the TensorShape/FixedRankTensorShape/TensorSpec classes. (#906)
These were added a bit of premature generalization which forced a bunch
of complexity into the stack.  In practice, only the CompactTensorShape
class is used: kernels that want to get the dimensions as a fixed array
or as a smallvector can use a method on it to project into those existing
types.

This patch just removes the classes we aren't using, the next patch will
rename CompactTensorShape to TensorShape.

MODULAR_ORIG_COMMIT_REV_ID: dd6662d81733cfa630c4dfa5514f0029d5a85bec