[KGEN] Introduce a kgen.param.value operation. (#944)
This operation projects things from the metaprogramming domain of
parameters into the SSA value domain of the runtime program.  Use
this as an opportunity refactor the printing and parsing logic for
parameters to be unified and nicer when in known contexts.

MODULAR_ORIG_COMMIT_REV_ID: 5fa5a7c5d412d63aac8c21ed4fb491306644f406