[KGEN] Switch the default parameter type to `index`. (#984)
This composes better with projections into SSA values when turning
a parameter into (e.g.) a stride to index with scf.for.  This also
leads to more architecture independent kernels in general.

There is a concern with making sure the math in the evaluator correctly
handles overflow.  When we build that, we should make sure the overflow
checks are done in the pointer size of the target, and that traps are
treated as kernel generation failures.

MODULAR_ORIG_COMMIT_REV_ID: 73f34c64d5e63a92e937952c296bdfc875bc743f