[KGEN] Create a new "meta" dialect and move !kgen.scalar into it. (#1000)
This provides a better layering between the kgen dialect which is
the "builtin stuff" that the compiler is hard coded to, and helpful
types and operators that are commonly used - more of a standard
library.

This fixes #978.

MODULAR_ORIG_COMMIT_REV_ID: f6c09df92919f1e78cc971b72f83c2cd7ac83ce2