[KGEN] Refactor parameter verification into a class.  NFC. (#1002)
The verification interface is just a single method, but it has some
state, and is soon to have mutually recursive implementation details.
Put the state in a class and move the implementation logic to methods
to help make it easiser to work with.

MODULAR_ORIG_COMMIT_REV_ID: b3fa226ea3495c88e9c07aa1b622db1db9ef6f97