[KGEN] Rename kgen.generate -> kgen.call, make it work with kgen.kernel. (#1038)
kgen.generate was the wrong name for this in the first place (my fault)
in that it doesn't cause generation of the generator, it invokes something
produced by the generator.  Renaming it to kgen.call makes this more clear
and also make it ok for it to work with kgen.kernel as well.

MODULAR_ORIG_COMMIT_REV_ID: b5ee2d674681bbe8c5b19e6092d1694b3e9c7d1b