[KGEN] Implement the start of a kgen.generator.interface op. (#1118)
This allows defining the signature for a family of kernel generators
that all conform to a signature.  So far we allow it to be called, but
don't support kernels and generators that conform to the interface.

MODULAR_ORIG_COMMIT_REV_ID: 74d92acca1fda80ebfa305d3b87ca7c0e997ba58