[KGEN] Expose verifyDeclMatchesInterface and use in GenerateKernels. (#1264)
This makes the diagnostics for mismatching generator interfaces much
more specific and precise.

This patch does have a problem though.  We're generating a note in the
library file saying "this is where the declaration was" which I'm not
sure how to make cooperate with -verify-diagnostics.  Because of this,
I had to change the run line of kgen-generate-errors.mlir switch to
"not kgen-generate" which effectively neuters the test.  I'd love advice
on how to fix this.

MODULAR_ORIG_COMMIT_REV_ID: 785d56c8dddf4a094b2cd05645915b2e70cf99f1