[KGEN] Enhance verification to check for cycles in the parameter graph. (#1286)
This builds a graph projection of the parameter def/use graph that
LLVM's GraphTraits can use, and then uses SCCIterator to walk it. When
we find a cyclic region, we put in some effort to diagnose the problem
in a clear way.

This fixes #938

MODULAR_ORIG_COMMIT_REV_ID: c76d241bff3f629072cfd59385afef62edbc7e77