[KGEN] Start producing specializations of generators for their input parameters. (#1341)
This substitutes the available input parameters into the body of a
cloned version of the generator, transforming a kgen.generator into
a kgen.kernel.  This is just a step, there are still a bunch of things
to do for this to be "real".

MODULAR_ORIG_COMMIT_REV_ID: 42cd36163c539e826f50d34f99bacb240471a8b7