[Support] Rename `TensorElementType` to `DType`, NFC. (#1371)
This class is more general than tensors, already being used by BEF
and KGEN, for example.  I chose to name this "DType" instead of
"DataType" because DataType is a very generic name, and "dtype" is
the colloquially used name in the ML community for this specific
abstraction.  This is also pretty pervasive, so a shorter name
makes code slightly easier to read.

This is the second third of Issue #1066, need to rename DataType.h
to DType.h, sigh.

MODULAR_ORIG_COMMIT_REV_ID: 91adfb2f1d08ef420a485c81a03491863c0e5560