[ErrorOr] Introduce macros to reduce error propagation boilerplate. NFC (#1391)
This introduces RETURN_ERROR and UNWRAP_ERROR macros, the first of which
propagates an error value out of the current scope and ignores the returned
normal value, and the second propagates error value out of scope, and binds
a normal value to a named local variable.

A cleaner way to do this would be to use GNU statement expressions to merge
these into one expression-like thing, but MSVC++ doesn't support them. :-(

I'm curious what folks think of the macro names and whether we should add
a M_ prefix to the macros to avoid name collisions.  I chose not to add
the prefix because it adds verbosity, but also because we can add it down
the line if there is a conflict in practice.

MODULAR_ORIG_COMMIT_REV_ID: 72da7de934e17d4dd63a4f9f16608a918e1567ff