[KGEN] Implement support for constraint checking in elaboration. (#1437)
This refactors ParameterRewriter to split a new ParameterEvaluator
class out that evaluates parameter expressions without rewriting a
kernel body.  This is used to prune expansions when constraints are
not satisfied.

One gap in this is that this doesn't handle the case when a kernel
/cannot/ be expanded.  This will need a revamp of the diagnostics
system and some other work that is still to come.

MODULAR_ORIG_COMMIT_REV_ID: cecc4a89f60673f1ff4ba6ed0bcc11fa0cb41f63