[KGEN] Delete dead kernels when elaboration fails. (#1471)
Elaboration of a kernel can fail for a number of reasons, including
that the instantiated IR doesn't verify.  When this happens we capture
an IR-independent error that can be reported to the user if necessary,
but we also must delete the unviable kernels.  This is not necessarily
a fatal error, other candidates may have been viable.

MODULAR_ORIG_COMMIT_REV_ID: 18c042c5ad04bba40809543874ee9833d0b32b26