[KGEN] Delete dead kernels when elaboration fails. (#1476)
Elaboration of a kernel can fail for a number of reasons, including
that the instantiated IR doesn't verify. When this happens we capture
an IR-independent error that can be reported to the user if necessary,
but we also must delete the unviable kernels. This is not necessarily
a fatal error, other candidates may have been viable.

This reapplies PR1471 with a fix for a use-after-free bug.  Thx to
Aman for reverting the bad patch!

MODULAR_ORIG_COMMIT_REV_ID: 02a98d512003ce60cc3639f2138a3df051532897