[KGEN] Expand generator/inputParam sets consistently within a kernel. (#1497)
A generated kernel may have generated sub-computations that are used
multiple times.  There may be many ways to expand these subcomputations,
but there should be an optimal one for any given environment.

Given this, it doesn't make sense to expand all permutations of the
generator within a final generated kernel: we now keep track of the
bindings used in a kernel in a new ElaboratedKernel structure, and
check to make sure future bindings are kept consistent with the
previous bindings.

This massively prunes the expansion and search space of generated
kernels in practice, but this is still a fundamentally expontential
algorithm.

MODULAR_ORIG_COMMIT_REV_ID: b84b21efdee6727f11ee79a0f428ba234242351e