[LLCL] Change waiter list to hold homogenous unique_function instances. (#1509)
This walks back a previous optimization for waiters that are too big
to fit in the inline representation of llvm::unique_function.  We
used heterogenous linked list nodes since use of malloc for each node
allowed each allocation to be the same size.

This was a good optimization, but doing one allocation for every node
is not ideal for other reasons.  We're moving to a world where we want
each waiter to be the same size, so we can put it in the payload of
the async value or batch them up into a single allocation.  As such,
remove the previous optimization and simplify the code to make way
for that.

MODULAR_ORIG_COMMIT_REV_ID: c0bf06e0671c58388bd39c2bf57a9049f79edeaf