[LLCL] Improve support for mutable lambdas. NFC. (#1561)
andThen invokes its lambda exactly once, so it is convenient for the
body of the lambda to move values out of the lambda capture list.  This
can improve performance etc.  This requires that the wrappers around
andThen preserve mutability though.  This will be used in a forthcoming
patch.

MODULAR_ORIG_COMMIT_REV_ID: 9c9ddfb0647969fc561097619216576ad217a5d9