[SoftMax] Rework softmax impls to avoid using "construct". (#1562)
These were the only uses of the `parallelForEachNMarkReady` API
which is all about making the AsyncValue "construct" pattern more
ergonomic.  Remove it.

This reworks both implementations a bit to use a different pattern
that explicitly manages a "work in progress" value that gets stored
into the asyncvalue at the end with emplace.  Along the way, this
shaves off a few refcount bumps and other captures.

test.vec.add.async4 also used this, but it was just a redundant
algorithm to show off this API, so just remove it.

MODULAR_ORIG_COMMIT_REV_ID: 74652ceb6082dcf457f741b9b0364455595d2e2a