[LLCL] Microoptimize AsyncValue transition to WaiterPresent state. (#1630)
We have a few-cycle window where we are initializing the inline
waiter in an AsyncValue (the first waiter added to it) and then
we change the state to ...WaiterPresent.  Use knowledge of the
bit encoding to switch this second change from a compare/xchg
loop to an atomic increment.  This is more efficient in terms of
instructions executed and hardware costs.

This unfortunately required changing the std::atomic to have
integer type, instead of being able to use atomic<WaitersAndState>
directly.  As a consequence, this patch introduces a bunch of
helper APIs in AsyncValue to paper over this difference.  One nice
thing about this is that this keeps all the memory consistency
decisions consistent.

Oen thing I'm not sure of is what memory consistency is required
for this - I went with the most pessimistic, but I'm sure this
can be relaxed by someone who knows consistency policies better
than me.

MODULAR_ORIG_COMMIT_REV_ID: 2201f073032b3ae5fe3ebd2b9520f6d36283f143