[Support/GraphRT] Fix a bunch of GCC warnings on the benchmark builder. (#1648)
GCC means well, but it is using limited dataflow to generate
diagnostics and therefore generates a bunch of false positives.  This
is well known to not be the right way to do things if you care about
QoI, but GCC is special and we want to work well with it.

The parenthesied VARIABLE thing was apparently added to appease
clang-tidy, so if that is important, we may need to silence the
warning with a pragma as well.

MODULAR_ORIG_COMMIT_REV_ID: 5c49be469deacbc2c6afbf82b4815cfee668b3d9