[Warnings] Two minor tweaks.  NFC. (#1651)
The warning on the GCC builder about free is a dataflow warning,
so you can't just silence it in the function with the problem: in
the case of ErrorOr, the methods get inlined and we'd have to silence
it in all there users of ErrorOr.  This isn't practical so just
silence it for all translation units that pull in ErrorOr.h.

Fix a previous warning silencing in AsyncValue.h that disabled a
static_assert.

MODULAR_ORIG_COMMIT_REV_ID: bcd7117b7d086bbdf0eebac6347a50b7cda214d1