[LLCL] Loosen an AsyncValue assertion. (#1665)
I added this as a defensive measure against what would otherwise
be a memory leak.  It turns out that IndirectAsyncValue's can be
destroyed before they are fulfilled, and this is happening on
the builder.

This needs a proper solution, but just relax the assertion for now
to unbreak the build.

MODULAR_ORIG_COMMIT_REV_ID: 40ad77eada9dbfaa849012fe890a3428e374ddbb