[LLCL] Fix a nasty bug in IndirectAsyncValue. (#1696)
In GraphRT (at the bottom of RegisterInfo::getOrCreateIndirectValue)
there is a case where we construct an IndirectAsyncValue and then
immediately destroy it.  This happens when another thread beat the
current thread to set a register value.  In that case, we immediately
destroy the IndirectAsyncValue.

This is all fine, but my inline-waiter optimization patch had a bug:
when in the unconstructed state, the 'value' RCRef won't be
initialized, so we shouldn't destroy it again.  This was manifesting
as breakage on the builder and ASAN failures.

This fixes #1666

MODULAR_ORIG_COMMIT_REV_ID: 8d7d97fe6ac75c13345a0f968f92193207bd7e20