[LLCL/TPWQ] Reimplement the await() algorithm. (#1815)
This changes how the allValuesDone semaphore is used: instead of
waiting on it an arbitrary number of times whenever we run out of
work, we wait on it exactly once, and signal it exactly once.  This
ensures the 'andThen'ing threads complete and don't have dangling
references to our stack, and eliminates a window of concurrency
bug that I'm sure is there but can't provoke easily with a test
case.

Note that this still has the "TODO: This code has a problem" issue.
When the "await()" thread runs out of patience and blocks on the
semaphore, it is depriving the system from that core for all other
work.  This can cause significant performance issues, and even
chained MNIST is hitting this.

I believe that the busywait logic is showing improvements because
it is deferring this from happening, but the right thing would be
to fix the core issue.

MODULAR_ORIG_COMMIT_REV_ID: ed81b8c95392e8e52c4afdf023d83bcf7e86fb1d