[MLIRToBEF] introduce ConcatenationTree to stop copying data around, NFC. (#1818)
This introduces a new "ConcatenationTree" ADT to Support which builds a
tree of data that can be emitted in arbitrary orders and then concatenated.
This is useful for BEF emission because it wants to generate various
sections in parallel, then plop them together into a logical output file.
The challenge before is that it iteratively does this, and this could cause
a lot of memory copies when dealing with large BEF files.

ConcatenationTree solves this by building a simple tree structure which
never has to copy the data around.  After constructing the tree of data,
the client can just traverse the data to emit it to an ostream, or
linearize it into an llvm::MemoryBuffer depending on their use-case.

This Fixes #1759.

MODULAR_ORIG_COMMIT_REV_ID: b0379ca1d938b86d4f559d1f8c2304f2adc523dd