[KGEN tests] Refactor and tidy up library. (#1823)
Several changes in this patch:

This splits out the unary_add test thing to a "test library" which
is there for testing the infra but isn't reasonable or interesting
otherwise.

This renames scalar_add to add to be more consistent with other
generators in this file, and renames the type parameter to type.

This adds comment flyer's about each generator.

This adds type constraints to various generators.

This renames constant_float to float_constant since the sense of it
is that it is useful for floating point literals, but doesn't
necessarily return a float type (for example it could produce
complex float or something).  integer_constant in
the future will also be compatible with FP types for example.

This adds a generator interface for float_constant, and implements
it correctly for f32, and adds constraints.

MODULAR_ORIG_COMMIT_REV_ID: b4c75374447de000d27f5191cabf30db7a86af8a