[LLCL] Stub out the build work for a new experimental WorkQueue. (#1869)
This will evolve over time and isn't default for any configuration,
but this allows playing with it and developing it in tree.  The
name is arbitrary and I'd love better suggestions, the major
compelling argument for this name is it allowed defining a
kNINE 🐶 enumerator.

MODULAR_ORIG_COMMIT_REV_ID: 8a978c7e64e2f866cf74c59922e5a89e32ba3802