[LLCL] Improve parallelForEachN() efficiency (#1917)
This is the (frowned upon but) commonly used version that blocks in
await.  Because it is going to donate a thread to run tasks when it
calls await, it makes sense for this to directly invoke the last
work item in the set that needs execution.  This allows faster pickup
and avoids threads going to sleep (in the absense of busy wait).

This Fixes #1636

MODULAR_ORIG_COMMIT_REV_ID: ed908946bc0f84ff97f70c78cda93dce55cf7ba7