[LLCL] Add support for over-aligned workqueues. (#1949)
This changes WorkQueue to be cache line size aligned, which enables
ThreadPoolWorkQueue have aligned members inline instead of indirect
(e.g. its RingBuffer). The NINE workqueue is already depending on
this, and ASAN is (correctly balking) about it.  This also adds a test for
the NINE workqueue.

It took me awhile to understand this, but the issue is apparently
that std::unique_ptr does support invoking the aligned deleter
when T is overaligned. The problem is that it allows a conversion
from std::unique_ptr to std::unique_ptr even though
Derived requires the aligned deleter and Base does not. This seems
like a bug in the C++ standard or libc++.

MODULAR_ORIG_COMMIT_REV_ID: e3a9752d2b1d1197d8391bd50b5d3ef2ad2fecc0