[LLCL] Default busy wait to 1ms. (#1961)
We know that zero is the wrong number, and I keep seeing graphs with
it left at the default.  We don't know if the right number is 1, 2
or something else, but 1 is a lot closer to a useful number than 2.

MODULAR_ORIG_COMMIT_REV_ID: 86a64f8b816d4253781b68195c1e41b09abbe9b8