[KGEN] add DenseArrayAttr to the allowlist for LLVM dialect. (#2000)
[KGEN] add DenseArrayAttr to the allowlist for LLVM dialect.

DenseArrayAttr is new and is used by the LLVM dialect.  it doesn't
have parameters so we don't need to rewrite it.  We need to add it
to the allowlist since it doesn't conform to
SubElementAttrInterface.

MODULAR_ORIG_COMMIT_REV_ID: 88e6ec4d00053d99e4f7b23070f77f159cb27e71