[KGEN] Introduce `in` and `in_dtype` operators (#2008)
* [KGEN] Introduce `in` and `in_dtype` operators

The `in` and `in_dtype` operators are variadic operators that check if
the value represented by the leftmost operand can be found in any of the
trailing right-hand-side operands.

* [kgen] Improve syntax for IN operators

The syntax for `in` and `in_dtype` have to changed to be

```
in(p1, [0, 1, 2])
in_dtype(type, [f32, f64])
```

For better readability.

* add test with two type parameters

* fold  with two operands to

MODULAR_ORIG_COMMIT_REV_ID: fccfbb7d42435c6427e2ff2f51703f51e543a381