[kgen] Add LLVM lowerings for buffer types and other ops (#2033)
This PR adds a conversions to LLVM for BufferType and SIMDType and associated ops. The former is converted to an LLVM struct with an i64 size and pointer to the data type. The latter is converted to the standard VectorType with one scalable dimension.

MODULAR_ORIG_COMMIT_REV_ID: e2f7fc7db406c3b570829769f43e8f8ffffa5b19