[KGEN] Expand generators/itf constraints to include a message. (#2084)
The constraint list on a generator is conceptually similar to a
static_assert compressed into the generator declaration, making
it more efficient for the elaborator to process.   That said,
we need to report elaboration errors in a nice way, so we should
allow/require the constraint to have a string explaining the
problem.

MODULAR_ORIG_COMMIT_REV_ID: 0fb9eabb6f3f64bcb12a54c66c90907596e9b41c