[KGEN] Add the task list as a MD file in the repository. (#2105)
This obsoletes the google doc, moving the source of truth
for implementation status into the codebase where it can be
kept up to date.

I took this as an opportunity to rearrange the doc so the
completed items are at the bottom of the doc for historical
purposes but the meat is at the top.

I'd love for the other task lists to get merged into this so
we have a single source of truth on this.

MODULAR_ORIG_COMMIT_REV_ID: dbef4221bb22de1a77a359f579cb30542259b838