[KGEN] Simplify the elaborators handling of cross-module symbol references. (#2204)
The elaborator now tracks which (possibly included) library a kernel
it is specializing came from.  This allows it to directly resolve calls
to it.

This also changes the cloning approach to be less eager.  It only needs
to clone over kernels that are cross module referenced.  This was a bug
even before we had #include support, but is showing up in the more
complex tests we have now.

MODULAR_ORIG_COMMIT_REV_ID: 1d682e40c65f61819e70c5589f43b36f8990e776