[KGEN] Improve the flexibility/power of `meta.buffer.cast`. (#2225)
It is important that buffer casts be allowed between things with
parameters, even if we cannot prove (yet) that they will be equal
to some constant size.  This will be later discovered by the
elaborator (which is prepared to handle discovering an error at
that level).

MODULAR_ORIG_COMMIT_REV_ID: 018fc5c25289a3a379cddc551962275761ee47ac