[KGEN] Revise storage of result parameters for generators/kernels/itfs. (#2227)
This moves the result parameter declarations to their own
`resultParamDecls` attribute that is separate from the `paramDecls`
attribute list maintaining the input parameters.

We used to maintain one list and an integer that indicated where in
the input the split between the two was.  The was a "clever" encoding,
but made a bunch of stuff more complicated, because the result
parameters of a generator are not in scope for its body.

This in turn caused problems when shadowing the result parameter with
the same name in the body of the generator.  This is illustrated by
the testcase, and is important/useful for an upcoming patch that causes
result parameters to get propagated out of calls.

MODULAR_ORIG_COMMIT_REV_ID: 318a583c9b2573bb36aa00c828079bfed27bc706