[kgen] Rename meta.buffer.cast to meta.buffer.rebind (#2288)
* [kgen] Rename meta.buffer.cast to meta.buffer.rebind

The name "cast" is confusing because it implies value conversions that
are not possible according to the op spec. The op is renamed to "rebind"
since it is binding the parameters of parameterized type to something
more or less specific.

* review comments

MODULAR_ORIG_COMMIT_REV_ID: bd635ad8014bff68f9bc9c7a74815fe5dfa56d9c