[KGEN] Build ConstraintSet's support for pointwise parameter bindings. (#2307)
This deconstructs and models parameter equality constraints of
parameters to constants.  This is enough to get us to diagnose
simple contradictions (as in the testcases), and nicely composes
with inferred constraints.

This need to get smarter in a few ways, including modeling of sets
of pointwise values, != constraints, and "x=y" constraints, and of
course relational constraints, but it is a start.

MODULAR_ORIG_COMMIT_REV_ID: 299e60db73f8d309b2bb5cff4f427ff5cd6542bb