[KGEN] Refactor ExecutionEngine to operate on KernelOps. (#2393)
Since the serialized format we want to save is the output of the elaborator, we should move the lower-to-llvm + emit object file to a self-contained unit. This also means that the API of the ExecutionEngine should not work on LLVM ops, it should work on KGEN ops.

MODULAR_ORIG_COMMIT_REV_ID: 579a969e040e64190b57e3ae2cc6d547a28838b0