[LLCL] `AsyncValue.h:` avoid moving from lvalues in `andThen` (#2398)
`andThen` takes a forwarding reference to a callable type for the waiter.  This
means it can take `lvalues` or `rvalues`. However, in one branch, there is an
unconditional use of `std::move` on the waiter.  This is not safe for `lvalues`.
`clang-tidy` warns about this with `bugprone-use-forwarding-reference` check.

Fix the problem for `lvalues` by using `std::forward` instead. This has the
correct behavior for the various value categories now.

MODULAR_ORIG_COMMIT_REV_ID: be0206b62fe7309464d431aad5446346a0acd61f