[KGEN] When building a KGEN library, you need to depend on the input … (#2420)
…source

Otherwise modifications to the mlir file will not trigger a rebuild of the target.

There is an argument to be made that we should also depend on the library include files, but I am not proficient in CMake enough to do that.

MODULAR_ORIG_COMMIT_REV_ID: a735696c3e1ce870fc51001e0660148e719c6d4c