[GraphRT] Rewrite the grt.repeat.i32 primitive. (#2488)
The previous behavior was to "unroll" the loop as soon as the condition
resolved, kicking off thousands of iterations at once without creating
any tasks.  While this can be "interesting" as a stress test, this is
not very composable or friendly for large bodies.

Solve this by changing the code to kick off each iteration of the loop
one at a time, after waiting on the previous iteration to complete.
This is also an imperfect approach (it reduces inter-iteration
parallelism) but is imperfect in a more conservative direction.

Because this implementation creates more tasks, it required adapting
some existing testcases.  the issue is that the bef-executor does an
await on the results of a testcase before kicking off the next test
case - in an attempt to get parallel testcases to not interlace their
output w.r.t each other.  This required us to actually return the
output chain from these.  Even that was not enough given how the
value tracking destructors run - the can (correctly) run at
unpredictable later time which is a problem given they have a side
effect of printing out stuff.  I tweaked things to dodge this, but
if more problems come up, we might disable this as being more trouble
than it is worth.

MODULAR_ORIG_COMMIT_REV_ID: add7be0f637a54c0cfa80242d9859269dec18730