[KGEN] Make the elaborator only work on generators/interfaces. (#2604)
Before this patch, the elaborator crunched on both kgen.kernel decls
as well as kgen.generator decls and lowered them into kgen.kernels.
This led to some semantic confusion: kgen.kernel could have a
parametric body before elaboration (but not after) and it makes it
less clear what phase the IR is in etc.

There is no reason to do this in the current design: instead, change
the elaborator to only pay attention to kgen.generator decls and
lower the roots that have no input parameters.  This gives us cleaner
contract and allows us to tighten up the verification logic for
kgen.kernel (which will be a follow-up patch).

Note that this patch has a really unfortunate side effect: as part
of elaboration of a top level generator, we need to introduce a new
kgen.kernel with the same name as the generator itself, but we can't
remove the existing generator first because it can be used by
multiple other things.  Avoid a collision (which would be autorenamed
to something_0 by SymbolTable) by appending a _kernel suffix.

This is suboptimal, I'd love it if someone has an idea here or if we
could add a method to the MLIR SymbolTable class that removes an op
from the symbol table without deleting the op (like its erase method
does).

This fixes #2581

MODULAR_ORIG_COMMIT_REV_ID: e6886479ebf64f9b444a4349b9164a27ba63dc81