[Kernels] Implement a vectorized elementWiseExp using KGEN (#2625)
This is one way to implement an elementWiseExp as a Kernel. The next step is to move all this logic into KGEN and implement a kernel that operates on buffers and performs all the simd-ization for us.

This is still good for validation and testing.

MODULAR_ORIG_COMMIT_REV_ID: 589e68f04e18cc17d54980204b947553e9b1bb67