[KGEN] Start introducing support for first class type parameters. (#2664)
We need the ability to write fully type generic algorithms, e.g. an
add which could be scalar or vector.  To do this, we need kgen
parameters that can themselves be arbitrary MLIR types.

This introduces a new 'kgen.type' type (written as 'type' in pretty
contexts) which is spelled MLIRTypeType and an attribute of that
type called MLIRTypeConstantAttr, along with associated plumbing to
allow it to get parsed, printed, and verified.

Rationale: The MLIR prefix isn't necessary, but I chose to go with
it to avoid the name "TypeType" which is correct but less specific
and will cause confusion down the line.  I chose to introduce
MLIRTypeConstantAttr because all parameter expressions have a type
and TypeAttr (correctly) does not.  This keeps "dtype" and "type"
naming and approach aligned.

This is a step towards implementing Issue #2504

MODULAR_ORIG_COMMIT_REV_ID: 6d455fbb843e9ad09a67f8c10b03241467a297f8