[KGEN] Introduce !kgen.paramtype to represent fully parameterized types. (#2665)
This is a parameterized type indicating that the elaborator needs to
resolve it to a concrete type when the parameter expression is resolved
to a concrete value.  Because it is a placeholder for a concrete type,
it auto-folds to a concrete type when one is already present.

MODULAR_ORIG_COMMIT_REV_ID: 5d1b16c05922dc1057acb50a5b006d798e16d94e