[KGEN] Split MLIRTypeConstantAttr into {Concrete|Parameterized}TypeConstantAttr (#2675)
It is valid to refer to types with type parameters before elaboration,
but they need to be gone after that.  Splitting this makes the core
predicate (isSimpleConstant) for determining this simple.

This is still WIP.  Notably, ParameterizedTypeConstantAttr::get
and ConcreteTypeConstantAttr::get do not yet enforce the correct
invariants.  That will come in the next patch.

MODULAR_ORIG_COMMIT_REV_ID: 5bea15c3a01d826840bef39c275a9409440648e6