[KGEN] Opaque wrappers take one struct for args and one for results. (#2677)
When calling into a JIT'ed function, we also don't statically know *how many* arguments are in the function. Therefore, we have to pack *all* the arguments into a single memory block. We represent this memory block with a packed struct. We already do the same with the results, and this will allow a simple integration with the Elaborator and the benchmark attributes by allowing us to pack user types into a single memory block.

MODULAR_ORIG_COMMIT_REV_ID: 58ae6c36d1df577305dc75af67b746565cd29a98