[KGEN] Generalize the histogram bucket tags. (#2678)
Histogram buckets can be tagged with anything - this will be especially
important for user-defined types. Because we want to support things like
that, we should allow the histogram to take any Attribute as a tag. In
the future we will provide this tag to types to have them generate
instances of themselves.

MODULAR_ORIG_COMMIT_REV_ID: aac6235dd480a260f4573a071f0358486bd7becd