[KGEN] Change compiled kernel abstraction, NFC. (#2685)
This adds an explicit abstraction for CompiledKernel that simplifies
some function signatures. The idea is that this way, you can invoke a
CompiledKernel exactly the way you would the original function to
simplify the mental model. Another benefit to this model is that you can
more easily add asserts to ensure the invoke signature matches the
kernel's signature (though that would be a fairly large patch that can
be added later).

Also adds a test for void kernels because voids are always tricky when
doing template magic.

MODULAR_ORIG_COMMIT_REV_ID: fa663f06bc28367e39dca3bee016996893426ba6