[KGEN] Rework EvalConfigurationAttr. (#2709)
EvalConfigurationAttr's previous design was to have a histogram of
configurations per-argument. This is nice from the standpoint of having
configurations map naturally to arguments, but is terrible because you
have to transpose a matrix that has a non-constant row dimension
(essentially) in order to get the list of all configurations to
evaluate.

This patch reworks the attributes to make it cleaner and simpler to
evaluate a generated kernel per-configuration in the elaborator.

MODULAR_ORIG_COMMIT_REV_ID: d636599bc06af7def8902f7379649e4c08f2de0d