[unit tests] Use lit to execute the unit tests  (#2700)
A mix of test executors is used in the codebase with `lit` being the
primary one for MLIR and regression tests. `CTest` was recently
introduced to
execute the new unit tests.  However, this has some drawbacks:

1. Different test executors increases developer mental burden for little
value add.
2. Additional tooling/infra would have to be built to plumb through
sanitizers, valgrind options, etc. for the unit tests. This does not
make sense to
build when it's already done for us with `lit`.

Instead, use `lit` to execute the unit tests by adding various
`lit.cfg.py` in the `unittests` directory for `Kernels` and `Support` to
specify the
`GoogleTest` as the `lit` format to use for test execution. Remove the
`CTest` machinery in `add_modular_unittest`.  Also, remove setting
`CTEST_OUTPUT_ON_FAILURE` environment variable in the startup scripts as
`CTest` is no longer used.

In the `lit.cfg.py` for the unittests in `Support` and `Kernels`,
explicitly turn `GTest` sharding off. When it is on by default, this
causes
malformed JSON files to be produced when using the Google Test formatter
with lit.

Fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: e7aa7956c6c353788fb8b06994d469bd95f5dfe8