[KGEN] Rework `add` and `lookup` in ExecutionEngine. (#2741)
This change fixes the behavior of `lookupOpaqueWrapper` - it has to use
the JITDylib of the original kernel, but lookup a different symbol.

This change also adds an option to `add` to allow the user to not
process all the kernels in the module. This is useful in the elaborator
because some of the kernels we aren't touching will have invalid
callees.

MODULAR_ORIG_COMMIT_REV_ID: 5e736f836152544b575131232afa53e82307f88a